♻️ Refactor shared ticket functions
